backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[hiredis]==5.2.1
pydantic[email]==2.10.4
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
alembic==1.14.1
snowflake-connector-python==3.17.0; python_version < "3.14"
litellm==1.55.4
pytest-asyncio==0.24.0
